requirements-dev.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
redis==5.0.3
fastapi-limiter==0.1.5
PyJWT[crypto]==2.8.0
testcontainers[postgresql]==3.7.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
python-multipart==0.0.9
PyYAML==6.0.1
ruff==0.4.3
